refactor(web): split settings-page.tsx into sections
settings-page.tsx: 1380 -> 885 lines. Extracted:
- lib/color-utils.ts (39) - RGBA parsing, palette utils
- settings/settings-inputs.tsx (194) - ColorInput, SizePicker, SettingRow
- settings/notifications-section.tsx (81) - push notification UI
- settings/key-capture.tsx (117) - keyboard shortcut capture
- settings/session-links-section.tsx (86) - inter-session links

voice-key.tsx updated to import formatKeyCode from new location.
SETTINGS array + DefaultSessionPicker remain inline (closure-bound
to dialog state). Pure structural move, no behavior changes.
