+++ web/src/components/settings-page.tsx
@@ -1,279 +1,21 @@
-import { Bell, BellOff, Cloud, Save } from 'lucide-react'
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { Save } from 'lucide-react'
+import { useEffect, useMemo, useRef, useState } from 'react'
 import { Dialog, DialogContent, DialogTitle } from '@/components/ui/dialog'
-import { getPushStatus, subscribeToPush, useSessionsStore, wsSend } from '@/hooks/use-sessions'
+import { useSessionsStore, wsSend } from '@/hooks/use-sessions'
 import { resolveToolDisplay, TOOL_DISPLAY_KEYS } from '@/lib/dashboard-prefs'
-import { clearCacheAndReload, cn } from '@/lib/utils'
+import { clearCacheAndReload } from '@/lib/utils'
 import { BUILD_VERSION } from '../../../src/shared/version'
-import { BUBBLE_COLOR_OPTIONS } from './transcript/group-view'
-
-// --- Color input with live preview ---
-const PALETTE = [
-  '#f9a8d4',
-  '#f472b6',
-  '#c084fc',
-  '#a78bfa',
-  '#818cf8',
-  '#60a5fa',
-  '#38bdf8',
-  '#22d3ee',
-  '#2dd4bf',
-  '#4ade80',
-  '#a3e635',
-  '#facc15',
-  '#fbbf24',
-  '#fb923c',
-  '#f87171',
-  '#e2e8f0',
-]
-
-const OPACITY_STEPS = [100, 85, 70, 50, 35, 20, 10, 0]
-
-function hexToRgba(hex: string, opacity: number): string {
-  const r = Number.parseInt(hex.slice(1, 3), 16)
-  const g = Number.parseInt(hex.slice(3, 5), 16)
-  const b = Number.parseInt(hex.slice(5, 7), 16)
-  return `rgba(${r},${g},${b},${(opacity / 100).toFixed(2)})`
-}
-
-function parseRgbaOpacity(rgba: string): number {
-  const m = rgba.match(/rgba?\(\s*\d+\s*,\s*\d+\s*,\s*\d+\s*,\s*([\d.]+)\s*\)/)
-  return m ? Math.round(Number.parseFloat(m[1]) * 100) : 100
-}
-
-function parseRgbaHex(rgba: string): string | null {
-  const m = rgba.match(/rgba?\(\s*(\d+)\s*,\s*(\d+)\s*,\s*(\d+)/)
-  if (!m) return null
-  const toHex = (n: number) => n.toString(16).padStart(2, '0')
-  return `#${toHex(Number(m[1]))}${toHex(Number(m[2]))}${toHex(Number(m[3]))}`
-}
-
-function ColorInput({
-  value,
-  onChange,
-  defaultColor,
-}: {
-  value: string
-  onChange: (v: string) => void
-  defaultColor: string
-}) {
-  const preview = value || defaultColor
-  const currentHex = (value && parseRgbaHex(value)) || null
-  const currentOpacity = value ? parseRgbaOpacity(value) : 100
-
-  function pickColor(hex: string) {
-    onChange(hexToRgba(hex, currentOpacity))
-  }
-
-  function pickOpacity(opacity: number) {
-    const hex = currentHex || parseRgbaHex(defaultColor) || PALETTE[0]
-    onChange(hexToRgba(hex, opacity))
-  }
-
-  return (
-    <div className="space-y-2">
-      <div className="flex flex-wrap gap-1">
-        {PALETTE.map(hex => (
-          <button
-            key={hex}
-            type="button"
-            onClick={() => pickColor(hex)}
-            className={`w-5 h-5 border transition-transform hover:scale-125 ${
-              currentHex === hex ? 'border-white scale-110' : 'border-border/50'
-            }`}
-            style={{ backgroundColor: hex }}
-            title={hex}
-          />
-        ))}
-      </div>
-      <div className="flex items-center gap-1">
-        <span className="text-[9px] text-muted-foreground w-8 shrink-0">alpha</span>
-        <div className="flex gap-0.5 flex-1">
-          {OPACITY_STEPS.map(op => (
-            <button
-              key={op}
-              type="button"
-              onClick={() => pickOpacity(op)}
-              className={`flex-1 h-5 text-[8px] font-mono border transition-colors ${
-                currentOpacity === op
-                  ? 'border-white text-foreground'
-                  : 'border-border/50 text-muted-foreground hover:border-border'
-              }`}
-              style={{ backgroundColor: hexToRgba(currentHex || parseRgbaHex(defaultColor) || PALETTE[0], op) }}
-            >
-              {op}
-            </button>
-          ))}
-        </div>
-      </div>
-      <div className="flex items-center gap-2">
-        <div className="w-6 h-6 border border-border shrink-0" style={{ backgroundColor: preview }} />
-        <span className="text-[10px] font-mono text-muted-foreground flex-1 truncate">{value || defaultColor}</span>
-        {value && (
-          <button
-            type="button"
-            onClick={() => onChange('')}
-            className="text-[9px] text-muted-foreground hover:text-foreground shrink-0 border border-border px-1.5 py-0.5"
-          >
-            reset
-          </button>
-        )}
-      </div>
-    </div>
-  )
-}
-
-const LABEL_SIZES = [
-  { id: 'xs', label: 'XS' },
-  { id: 'sm', label: 'S' },
-  { id: '', label: 'M' },
-  { id: 'lg', label: 'L' },
-  { id: 'xl', label: 'XL' },
-]
-
-function SizePicker({ value, onChange }: { value: string; onChange: (v: string) => void }) {
-  return (
-    <div className="flex gap-0.5">
-      {LABEL_SIZES.map(s => (
-        <button
-          key={s.id}
-          type="button"
-          onClick={() => onChange(s.id)}
-          className={`px-2 py-0.5 text-[9px] font-mono border transition-colors ${
-            value === s.id
-              ? 'border-white text-foreground bg-muted'
-              : 'border-border/50 text-muted-foreground hover:border-border'
-          }`}
-        >
-          {s.label}
-        </button>
-      ))}
-    </div>
-  )
-}
-
-// --- Cloud icon for server settings ---
-function ServerIcon() {
-  return (
-    <span title="Server setting (shared)">
-      <Cloud className="w-3 h-3 text-blue-400/70 shrink-0" />
-    </span>
-  )
-}
-
-// --- Setting row wrapper ---
-function SettingRow({
-  label,
-  description,
-  server,
-  children,
-}: {
-  label: string
-  description: string
-  server?: boolean
-  children: React.ReactNode
-}) {
-  return (
-    <div className="flex items-center justify-between gap-3">
-      <div className="flex items-start gap-1.5 min-w-0">
-        {server && <ServerIcon />}
-        <div className="min-w-0">
-          <div className="text-sm text-foreground">{label}</div>
-          <div className="text-[10px] text-muted-foreground">{description}</div>
-        </div>
-      </div>
-      <div className="shrink-0">{children}</div>
-    </div>
-  )
-}
-
-// --- Group header ---
-function GroupHeader({ label }: { label: string }) {
-  return (
-    <div className="text-[10px] font-bold text-muted-foreground uppercase tracking-wider pt-3 pb-1 border-t border-border first:border-t-0 first:pt-0">
-      {label}
-    </div>
-  )
-}
-
-// --- Notifications (inline, not a separate tab) ---
-function NotificationsSection() {
-  const [pushState, setPushState] = useState<
-    'loading' | 'unsupported' | 'prompt' | 'subscribing' | 'subscribed' | 'denied'
-  >('loading')
-
-  useEffect(() => {
-    getPushStatus().then(status => {
-      if (!status.supported) setPushState('unsupported')
-      else if (status.subscribed) setPushState('subscribed')
-      else if (status.permission === 'denied') setPushState('denied')
-      else setPushState('prompt')
-    })
-  }, [])
-
-  async function handlePushToggle() {
-    if (pushState === 'subscribing') return
-    setPushState('subscribing')
-    const result = await subscribeToPush()
-    setPushState(result.success ? 'subscribed' : 'denied')
-  }
-
-  async function handleReRegister() {
-    if (pushState === 'subscribing') return
-    setPushState('subscribing')
-    try {
-      const reg = await navigator.serviceWorker.getRegistration('/sw.js')
-      if (reg) {
-        const sub = await reg.pushManager.getSubscription()
-        if (sub) await sub.unsubscribe()
-      }
-    } catch {}
-    const result = await subscribeToPush()
-    setPushState(result.success ? 'subscribed' : 'denied')
-  }
-
-  return (
-    <div className="space-y-2">
-      <div className="flex items-center justify-between">
-        <div>
-          <div className="text-sm text-foreground">Push notifications</div>
-          <div className="text-[10px] text-muted-foreground">Get notified when Claude needs input</div>
-        </div>
-        <button
-          type="button"
-          onClick={handlePushToggle}
-          disabled={pushState === 'unsupported' || pushState === 'loading'}
-          className={`flex items-center gap-1.5 px-3 py-1.5 text-xs border transition-colors ${
-            pushState === 'subscribed'
-              ? 'bg-active/20 text-active border-active/50'
-              : pushState === 'denied'
-                ? 'bg-red-400/20 text-red-400 border-red-400/50'
-                : pushState === 'unsupported'
-                  ? 'bg-muted text-muted-foreground border-border cursor-not-allowed'
-                  : 'bg-transparent text-foreground border-border hover:border-primary'
-          }`}
-        >
-          {pushState === 'subscribed' ? <Bell className="w-3 h-3" /> : <BellOff className="w-3 h-3" />}
-          {pushState === 'loading' && '...'}
-          {pushState === 'unsupported' && 'Not supported'}
-          {pushState === 'subscribing' && 'Enabling...'}
-          {pushState === 'subscribed' && 'Enabled'}
-          {pushState === 'denied' && 'Denied'}
-          {pushState === 'prompt' && 'Enable'}
-        </button>
-      </div>
-      {pushState === 'subscribed' && (
-        <button
-          type="button"
-          onClick={handleReRegister}
-          className="text-[10px] text-muted-foreground hover:text-foreground underline underline-offset-2 transition-colors"
-        >
-          Re-register push (use after VAPID key change)
-        </button>
-      )}
-    </div>
-  )
-}
+import { KeyCapture } from './settings/key-capture'
+import { NotificationsSection } from './settings/notifications-section'
+import { ProjectLinksSection } from './settings/session-links-section'
+import {
+  BubbleColorPicker,
+  ColorInput,
+  GroupHeader,
+  ServerIcon,
+  SettingRow,
+  SizePicker,
+} from './settings/settings-inputs'
 
 // --- Default session picker ---
 function DefaultSessionPicker({ value, onChange }: { value: string; onChange: (v: string) => void }) {
@@ -1139,242 +881,5 @@ export function SettingsDialog({ open, onOpenChange }: { open: boolean; onOpenCh
         </div>
       </DialogContent>
     </Dialog>
-  )
-}
-
-// --- Bubble color picker ---
-
-const BUBBLE_PREVIEW_COLORS: Record<string, string> = {
-  blue: 'bg-[#2563eb]',
-  teal: 'bg-teal-600',
-  purple: 'bg-purple-600',
-  green: 'bg-emerald-600',
-  orange: 'bg-amber-600',
-  pink: 'bg-pink-600',
-  indigo: 'bg-indigo-600',
-}
-
-function BubbleColorPicker({ value, onChange }: { value: string; onChange: (color: string) => void }) {
-  return (
-    <div className="flex gap-1.5">
-      {BUBBLE_COLOR_OPTIONS.map(color => (
-        <button
-          key={color}
-          type="button"
-          onClick={() => onChange(color)}
-          className={cn(
-            'w-5 h-5 rounded-full transition-all',
-            BUBBLE_PREVIEW_COLORS[color],
-            value === color
-              ? 'ring-2 ring-white ring-offset-1 ring-offset-background scale-110'
-              : 'opacity-70 hover:opacity-100',
-          )}
-          title={color}
-        />
-      ))}
-    </div>
-  )
-}
-
-// --- Key capture for push-to-talk ---
-
-const CODE_LABELS: Record<string, string> = {
-  ShiftLeft: 'Left Shift',
-  ShiftRight: 'Right Shift',
-  ControlLeft: 'Left Ctrl',
-  ControlRight: 'Right Ctrl',
-  AltLeft: 'Left Alt',
-  AltRight: 'Right Alt',
-  MetaLeft: 'Left Cmd',
-  MetaRight: 'Right Cmd',
-  ArrowUp: 'Up',
-  ArrowDown: 'Down',
-  ArrowLeft: 'Left',
-  ArrowRight: 'Right',
-  Backspace: 'Backspace',
-  Delete: 'Del',
-  Insert: 'Ins',
-  Home: 'Home',
-  End: 'End',
-  PageUp: 'PgUp',
-  PageDown: 'PgDn',
-  PrintScreen: 'PrtSc',
-  ScrollLock: 'ScrLk',
-  Pause: 'Pause',
-  NumLock: 'NumLk',
-  ContextMenu: 'Menu',
-  Space: 'Space',
-  CapsLock: 'Caps',
-  Enter: 'Enter',
-  Tab: 'Tab',
-}
-
-export function formatKeyCode(code: string): string {
-  if (code in CODE_LABELS) return CODE_LABELS[code]
-  const fKey = code.match(/^F(\d{1,2})$/)
-  if (fKey) return `F${fKey[1]}`
-  const numpad = code.match(/^Numpad(.+)$/)
-  if (numpad) return `Num ${numpad[1]}`
-  const letter = code.match(/^Key([A-Z])$/)
-  if (letter) return letter[1]
-  const digit = code.match(/^Digit(\d)$/)
-  if (digit) return digit[1]
-  return code
-}
-
-// Disallowed: keys that conflict with normal usage
-const DISALLOWED_KEYS = new Set(['Escape', 'Tab', 'Enter', 'Backspace', 'Delete'])
-
-function KeyCapture({ value, onChange }: { value: string | null; onChange: (code: string | null) => void }) {
-  const [capturing, setCapturing] = useState(false)
-  const buttonRef = useRef<HTMLButtonElement>(null)
-
-  const handleCapture = useCallback(() => setCapturing(true), [])
-
-  useEffect(() => {
-    if (!capturing) return
-
-    function handleKeyDown(e: KeyboardEvent) {
-      e.preventDefault()
-      e.stopPropagation()
-      if (e.code === 'Escape') {
-        setCapturing(false)
-        return
-      }
-      if (DISALLOWED_KEYS.has(e.code)) return
-      onChange(e.code)
-      setCapturing(false)
-    }
-
-    function handleClickOutside(e: MouseEvent) {
-      if (!buttonRef.current?.contains(e.target as Node)) setCapturing(false)
-    }
-
-    // Small delay so the click that opened capture doesn't immediately fire
-    const t = setTimeout(() => {
-      window.addEventListener('keydown', handleKeyDown, { capture: true })
-      window.addEventListener('mousedown', handleClickOutside)
-    }, 100)
-
-    return () => {
-      clearTimeout(t)
-      window.removeEventListener('keydown', handleKeyDown, { capture: true })
-      window.removeEventListener('mousedown', handleClickOutside)
-    }
-  }, [capturing, onChange])
-
-  return (
-    <div className="flex items-center gap-2">
-      <button
-        ref={buttonRef}
-        type="button"
-        onClick={capturing ? () => setCapturing(false) : handleCapture}
-        className={cn(
-          'px-3 py-1 text-xs font-mono border rounded transition-all min-w-[100px] text-center',
-          capturing
-            ? 'border-blue-500 bg-blue-500/20 text-blue-400 animate-pulse'
-            : value
-              ? 'border-border bg-muted text-foreground'
-              : 'border-border/50 bg-muted/50 text-muted-foreground',
-        )}
-      >
-        {capturing ? 'Press a key...' : value ? formatKeyCode(value) : 'Not set'}
-      </button>
-      {value && !capturing && (
-        <button
-          type="button"
-          onClick={() => onChange(null)}
-          className="text-[10px] text-muted-foreground hover:text-destructive"
-        >
-          clear
-        </button>
-      )}
-    </div>
-  )
-}
-
-// --- Project Links Management ---
-
-const LINKS_API = `${window.location.protocol}//${window.location.host}/api/links`
-
-interface LinkItem {
-  cwdA: string
-  cwdB: string
-  nameA: string
-  nameB: string
-  createdAt: number
-  lastUsed: number
-  online: boolean
-  sessionIdA?: string
-  sessionIdB?: string
-}
-
-function ProjectLinksSection() {
-  const [links, setLinks] = useState<LinkItem[]>([])
-  const [loading, setLoading] = useState(true)
-
-  const fetchLinks = useCallback(async () => {
-    try {
-      const res = await fetch(LINKS_API)
-      if (!res.ok) return
-      const data = (await res.json()) as { links: LinkItem[] }
-      setLinks(data.links)
-    } catch {
-      // network error
-    } finally {
-      setLoading(false)
-    }
-  }, [])
-
-  useEffect(() => {
-    fetchLinks()
-  }, [fetchLinks])
-
-  async function removeLink(cwdA: string, cwdB: string) {
-    await fetch(LINKS_API, {
-      method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ cwdA, cwdB }),
-    })
-    fetchLinks()
-  }
-
-  function formatAge(ts: number): string {
-    const diff = Date.now() - ts
-    if (diff < 60_000) return 'just now'
-    if (diff < 3600_000) return `${Math.floor(diff / 60_000)}m ago`
-    if (diff < 86400_000) return `${Math.floor(diff / 3600_000)}h ago`
-    return `${Math.floor(diff / 86400_000)}d ago`
-  }
-
-  if (loading) return <div className="text-xs text-muted-foreground font-mono">Loading...</div>
-
-  if (links.length === 0) {
-    return (
-      <div className="text-xs text-muted-foreground font-mono py-2">
-        No project links. Links are created when you approve inter-project messaging.
-      </div>
-    )
-  }
-
-  return (
-    <div className="space-y-2">
-      {links.map(link => (
-        <div key={`${link.cwdA}:${link.cwdB}`} className="flex items-center gap-2 text-xs">
-          <span className={cn('w-2 h-2 rounded-full shrink-0', link.online ? 'bg-green-400' : 'bg-zinc-600')} />
-          <span className="text-teal-400 font-mono truncate">{link.nameA}</span>
-          <span className="text-muted-foreground">↔</span>
-          <span className="text-sky-400 font-mono truncate">{link.nameB}</span>
-          <span className="text-[9px] text-muted-foreground/50 ml-auto shrink-0">{formatAge(link.lastUsed)}</span>
-          <button
-            type="button"
-            onClick={() => removeLink(link.cwdA, link.cwdB)}
-            className="text-[9px] text-muted-foreground hover:text-destructive shrink-0"
-          >
-            x
-          </button>
-        </div>
-      ))}
-    </div>
   )
 }
